Remove unnecessary date formatting
The date formatting function for 'publishedAt' has been removed from both the 'card.js' and 'gallery/[id].js' files in the codebase. This change reflects that the formatted date was no longer being used or necessary in these components. The 'priority' attribute was also added to the 'Image' component in 'card.js' to optimise image loading.