Add SearchSelectInput component and implement it in Gallery
A new SearchSelectInput component has been created and implemented in the GalleryFilter component. The new component allows users to search and select from a list of options. This has replaced the previous select input elements for operator and aircraft type in the gallery filter. Functions for selecting an operator and an aircraft type have been added and used as callbacks for the new component.